refactor(bulk-action): move adapters into controllers; footer-button UI
Address PR #1513 review (Azgaar):
- Move each migrated menu's bulk logic (delete cascade, summary, predicates)
  out of the generic bulk-action/adapters/ fixture and into its own controller
  domain (states/cultures/religions/regiments/markets). The controller now
  assembles its BulkEntityAdapter inline and passes it to the BulkActionBar; only
  the generic seam stays under bulk-action/. Pure cascade logic remains unit-tested.
- Replace the full-width top toolbar with a single icon button in each menu's
  footer button row; toggling it reveals row checkboxes plus inline footer
  actions (delete/lock/unlock/color) and Select-all + count.

Legacy-JS adapters and the window.bulkBars bridge stay as-is (cleanup deferred
to the JS->TS migration, per review).
