+++ src/controllers/bulk-action/bulk-action-bar.ts
@@ -3,21 +3,22 @@ import type { BulkEntityAdapter } from "./bulk-entity-adapter";
 import { BulkSelection } from "./bulk-selection";
 
 /**
- * DOM glue for one list menu. Adds a "Bulk Options" toggle at the top-right of the
- * menu; opening it reveals per-row checkboxes plus an inline toolbar (Select all +
- * Delete / Lock / Unlock / Set color, as the adapter supports). The toolbar stays
- * open across actions and the selection is kept (deleted rows simply drop off), so
- * actions can be chained. Wires DOM events to a shared BulkSelection and the menu's
- * BulkEntityAdapter, and re-syncs after the list re-renders. Attaches to both
+ * DOM glue for one list menu. Adds a single icon button to the menu's footer button
+ * row (like the other footer buttons). Toggling it on reveals per-row checkboxes plus
+ * a compact set of inline footer controls — Select all, an "N selected" count, and
+ * Delete / Lock / Unlock / Set color icon buttons (only those the adapter supports).
+ * The controls stay open across actions and the selection is kept (deleted rows simply
+ * drop off), so actions can be chained. Wires DOM events to a shared BulkSelection and
+ * the menu's BulkEntityAdapter, and re-syncs after the list re-renders. Attaches to both
  * legacy-JS and migrated-TS lists, which share the same `.table` of row `<div>`s.
  */
 export class BulkActionBar {
   private readonly adapter: BulkEntityAdapter;
   private readonly selection: BulkSelection;
   private container: HTMLElement | null = null;
-  private bar: HTMLElement | null = null;
+  private footer: HTMLElement | null = null;
   private toggle: HTMLButtonElement | null = null;
-  private toolbar: HTMLElement | null = null;
+  private inline: HTMLElement | null = null;
   private selectAllCheckbox: HTMLInputElement | null = null;
   private countLabel: HTMLElement | null = null;
   private bulkMode = false;
@@ -27,11 +28,12 @@ export class BulkActionBar {
     this.selection = new BulkSelection(id => adapter.isDeletable(id));
   }
 
-  /** Create the bar (idempotent) and wire events. Call once the container exists. */
+  /** Create the footer controls (idempotent) and wire events. Call once the menu exists. */
   mount(): void {
     this.container = document.getElementById(this.adapter.containerId);
-    if (!this.container) return;
-    if (this.bar) {
+    this.footer = document.getElementById(this.adapter.footerId);
+    if (!this.container || !this.footer) return;
+    if (this.toggle) {
       this.sync();
       return;
     }
@@ -40,43 +42,49 @@ export class BulkActionBar {
     const canColor = this.adapter.supportsColor && typeof this.adapter.setColor === "function";
     const lockButtons = canLock
       ? /* html */ `
-        <button type="button" class="bulkLock" data-tip="Lock selected rows (protects from regeneration and bulk delete)">Lock</button>
-        <button type="button" class="bulkUnlock" data-tip="Unlock selected rows">Unlock</button>`
+        <button type="button" class="bulkLock icon-lock" data-tip="Lock selected rows (protects from regeneration and bulk delete)"></button>
+        <button type="button" class="bulkUnlock icon-lock-open" data-tip="Unlock selected rows"></button>`
       : "";
     const colorButton = canColor
-      ? /* html */ `<button type="button" class="bulkColor" data-tip="Set color of selected rows">Set color</button>`
+      ? /* html */ `<button type="button" class="bulkColor icon-brush" data-tip="Set color of selected rows"></button>`
       : "";
 
-    const bar = document.createElement("div");
-    bar.className = "bulkActionBar";
-    bar.innerHTML = /* html */ `
-      <button type="button" class="bulkOptionsToggle" data-tip="Show bulk selection options" aria-expanded="false">Bulk Options ▾</button>
-      <div class="bulkToolbar" hidden>
-        <label class="bulkSelectAll" data-tip="Select or deselect all visible rows">
-          <input type="checkbox" class="bulkSelectAllCheckbox native" /> Select all
-        </label>
-        <span class="bulkCount">0 selected</span>
-        <button type="button" class="bulkDelete" data-tip="Delete selected rows">Delete</button>
-        ${lockButtons}
-        ${colorButton}
-      </div>`;
+    // The single entry button — styled like the menu's other footer icon buttons.
+    const toggle = document.createElement("button");
+    toggle.type = "button";
+    toggle.className = "bulkToggle icon-check-empty";
+    toggle.dataset.tip = "Bulk select: pick multiple rows, then delete/lock/recolor at once";
+    toggle.setAttribute("aria-pressed", "false");
 
-    // place at the top of the menu (above the column header)
-    const host = this.container.closest(".dialog") ?? this.container.parentElement ?? this.container;
-    host.insertAdjacentElement("afterbegin", bar);
+    // The inline controls, hidden until bulk mode is on.
+    const inline = document.createElement("span");
+    inline.className = "bulkInline";
+    inline.hidden = true;
+    inline.innerHTML = /* html */ `
+      <label class="bulkSelectAll" data-tip="Select or deselect all visible rows">
+        <input type="checkbox" class="bulkSelectAllCheckbox native" /> All
+      </label>
+      <span class="bulkCount">0 selected</span>
+      <button type="button" class="bulkDelete icon-trash" data-tip="Delete selected rows"></button>
+      ${lockButtons}
+      ${colorButton}`;
 
-    this.bar = bar;
-    this.toggle = bar.querySelector(".bulkOptionsToggle");
-    this.toolbar = bar.querySelector(".bulkToolbar");
-    this.selectAllCheckbox = bar.querySelector(".bulkSelectAllCheckbox");
-    this.countLabel = bar.querySelector(".bulkCount");
+    // Place the controls before a trailing search box if there is one, else at the end.
+    const searchAnchor = this.footer.querySelector(":scope > label");
+    this.footer.insertBefore(toggle, searchAnchor);
+    this.footer.insertBefore(inline, searchAnchor);
 
-    this.toggle?.addEventListener("click", () => this.toggleBulkMode());
+    this.toggle = toggle;
+    this.inline = inline;
+    this.selectAllCheckbox = inline.querySelector(".bulkSelectAllCheckbox");
+    this.countLabel = inline.querySelector(".bulkCount");
+
+    toggle.addEventListener("click", () => this.toggleBulkMode());
     this.selectAllCheckbox?.addEventListener("change", () => this.onToggleSelectAll());
-    bar.querySelector(".bulkDelete")?.addEventListener("click", () => this.onDelete());
-    bar.querySelector(".bulkLock")?.addEventListener("click", () => this.onSetLock(true));
-    bar.querySelector(".bulkUnlock")?.addEventListener("click", () => this.onSetLock(false));
-    bar.querySelector(".bulkColor")?.addEventListener("click", () => this.onSetColor());
+    inline.querySelector(".bulkDelete")?.addEventListener("click", () => this.onDelete());
+    inline.querySelector(".bulkLock")?.addEventListener("click", () => this.onSetLock(true));
+    inline.querySelector(".bulkUnlock")?.addEventListener("click", () => this.onSetLock(false));
+    inline.querySelector(".bulkColor")?.addEventListener("click", () => this.onSetColor());
 
     // delegate per-row checkbox changes from the container
     this.container.addEventListener("change", event => {
@@ -94,7 +102,7 @@ export class BulkActionBar {
 
   /** Re-apply per-row checkboxes (only in bulk mode) after a list re-render. */
   sync(): void {
-    if (!this.container || !this.bar) return;
+    if (!this.container || !this.toggle) return;
     const rows = Array.from(this.container.children) as HTMLElement[];
     rows.forEach(row => {
       const id = this.adapter.getRowId(row);
@@ -121,10 +129,12 @@ export class BulkActionBar {
 
   private toggleBulkMode(): void {
     this.bulkMode = !this.bulkMode;
-    if (this.toolbar) this.toolbar.hidden = !this.bulkMode;
-    this.toggle?.setAttribute("aria-expanded", String(this.bulkMode));
+    if (this.inline) this.inline.hidden = !this.bulkMode;
+    this.toggle?.setAttribute("aria-pressed", String(this.bulkMode));
     this.toggle?.classList.toggle("active", this.bulkMode);
-    if (this.toggle) this.toggle.textContent = this.bulkMode ? "Bulk Options ▴" : "Bulk Options ▾";
+    this.toggle?.classList.toggle("icon-check-empty", !this.bulkMode);
+    this.toggle?.classList.toggle("icon-ok-squared", this.bulkMode);
+    if (!this.bulkMode) this.selection.clear();
     this.sync();
   }
 
@@ -192,7 +202,7 @@ export class BulkActionBar {
   }
 
   private updateBar(): void {
-    if (!this.bar) return;
+    if (!this.toggle) return;
     const count = this.selection.count;
     if (this.countLabel) this.countLabel.textContent = `${count} selected`;
